fixup! add macos keyboard support
