+++ lib/helpers/load-os-module.js
@@ -6,14 +6,15 @@
  */
 
 /**
+ * @param {string} name
  * @param {{[key: string]: DeferModule<T>}} deferredModules
  * @return {T}
  * @template T
  */
-function loadOsModule(deferredModules) {
+function loadOsModule(name, deferredModules) {
   const { platform } = process;
   if (!(platform in deferredModules)) {
-    throw new Error(`No module for platform "${platform}"`);
+    throw new Error(`No "${name}" module for platform "${platform}"`);
   }
 
   return deferredModules[platform]();
