+++ lib/helpers/macos/KeyCode.js
@@ -1,0 +1,107 @@
+'use strict';
+
+/**
+ * Map of key names to Apple System Events key codes.
+ * @enum {number}
+ * - https://eastmanreference.com/complete-list-of-applescript-key-codes
+ * - https://github.com/guidepup/guidepup/blob/main/src/macOS/KeyCodes.ts
+ * - https://manytricks.com/keycodes/
+ */
+const KeyCode = exports.KeyCode = {
+  a: 0,
+  s: 1,
+  d: 2,
+  f: 3,
+  h: 4,
+  g: 5,
+  z: 6,
+  x: 7,
+  c: 8,
+  v: 9,
+  b: 11,
+  q: 12,
+  w: 13,
+  e: 14,
+  r: 15,
+  y: 16,
+  t: 17,
+  1: 18,
+  2: 19,
+  3: 20,
+  4: 21,
+  6: 22,
+  5: 23,
+  equal: 24,
+  9: 25,
+  7: 26,
+  minus: 27,
+  8: 28,
+  0: 29,
+  squareRight: 30,
+  o: 31,
+  u: 32,
+  squareLeft: 33,
+  i: 34,
+  p: 35,
+  return: 36,
+  l: 37,
+  j: 38,
+  quote: 39,
+  k: 40,
+  semicolon: 41,
+  backslash: 42,
+  comma: 43,
+  slash: 44,
+  n: 45,
+  m: 46,
+  period: 47,
+  tab: 48,
+  space: 49,
+  backtick: 50,
+  backspace: 51,
+  escape: 53,
+  command: 55,
+  shift: 57,
+  option: 58,
+  control: 59,
+  fn: 63,
+  numpadDecimal: 65,
+  numpadMultiply: 67,
+  numpadAdd: 69,
+  numpadClear: 71,
+  numpadDivide: 75,
+  numpadSubtract: 78,
+  numpadEqual: 81,
+  numpad0: 82,
+  numpad1: 83,
+  numpad2: 84,
+  numpad3: 85,
+  numpad4: 86,
+  numpad5: 87,
+  numpad6: 88,
+  numpad7: 89,
+  numpad8: 91,
+  numpad9: 92,
+  f5: 96,
+  f6: 97,
+  f7: 98,
+  f3: 99,
+  f8: 100,
+  f9: 101,
+  f11: 103,
+  f10: 109,
+  f12: 111,
+  help: 114,
+  home: 115,
+  pageUp: 116,
+  delete: 117,
+  f4: 118,
+  end: 119,
+  f2: 120,
+  pageDown: 121,
+  f1: 122,
+  arrowLeft: 123,
+  arrowRight: 124,
+  arrowDown: 125,
+  arrowUp: 126,
+};
