+++ lib/helpers/macos/KeyboardAction.js
@@ -1,0 +1,73 @@
+'use strict';
+
+/**
+ * Map of unicode characters to keyboard actions in System Events.
+ * @enum {string}
+ */
+// Commented out keys have known mapping in System Events.
+const KeyboardAction = exports.KeyboardAction = {
+  // '\ue000': 'unidentified'
+  // '\ue001': 'cancel',
+  '\ue002': 'help',
+  '\ue003': 'backspace',
+  '\ue004': 'tab',
+  // '\ue005': 'clear',
+  '\ue006': 'return',
+  '\ue007': 'enter',
+  '\ue008': 'shift',
+  '\ue009': 'control',
+  '\ue00a': 'option',
+  '\ue00b': 'pause',
+  '\ue00c': 'escape',
+  '\ue00d': 'space',
+  '\ue00e': 'pageUp',
+  '\ue00f': 'pageDown',
+  '\ue010': 'end',
+  '\ue011': 'home',
+  '\ue012': 'arrowLeft',
+  '\ue013': 'arrowUp',
+  '\ue014': 'arrowRight',
+  '\ue015': 'arrowDown',
+  // '\ue016': 'insert',
+  '\ue017': 'delete',
+  '\ue018': 'semicolon',
+  '\ue019': 'equal',
+
+  '\ue01a': 'numpad0',
+  '\ue01b': 'numpad1',
+  '\ue01c': 'numpad2',
+  '\ue01d': 'numpad3',
+  '\ue01e': 'numpad4',
+  '\ue01f': 'numpad5',
+  '\ue020': 'numpad6',
+  '\ue021': 'numpad7',
+  '\ue022': 'numpad8',
+  '\ue023': 'numpad9',
+
+  '\ue024': 'numpadMultiply',
+  '\ue025': 'numpadAdd',
+  // '\ue026': 'separator',
+  '\ue027': 'numpadSubtract',
+  '\ue028': 'numpadDecimal',
+  '\ue029': 'numpadDivide',
+
+  '\ue031': 'f1',
+  '\ue032': 'f2',
+  '\ue033': 'f3',
+  '\ue034': 'f4',
+  '\ue035': 'f5',
+  '\ue036': 'f6',
+  '\ue037': 'f7',
+  '\ue038': 'f8',
+  '\ue039': 'f9',
+  '\ue03a': 'f10',
+  '\ue03b': 'f11',
+  '\ue03c': 'f12',
+
+  '\ue03d': 'command',
+  '\ue040': 'backtick',
+
+  '\b': 'backspace',
+  '\t': 'tab',
+  ' ': 'space',
+};
