+++ lib/helpers/macos/Modifier.js
@@ -1,0 +1,13 @@
+'use strict';
+
+/**
+ * Map of modifier key names to System Events modifier identifiers.
+ * @enum {string}
+ */
+const Modifier = exports.Modifier = {
+  command: 'command',
+  control: 'control',
+  fn: 'fn',
+  option: 'option',
+  shift: 'shift',
+};
