+++ lib/modules/interaction.js
@@ -5,7 +5,7 @@
 const { loadOsModule } = require('../helpers/load-os-module');
 
 /** @type {ATDriverModules.Interaction} */
-module.exports = loadOsModule({
+module.exports = loadOsModule('interaction', {
   win32() {
     return require('./win32/interaction');
   },
