+++ lib/modules/macos/interaction.js
@@ -2,7 +2,8 @@
 
 'use strict';
 
-const { runScript, renderScript, parseCodePoints } = require('../../helpers/macos/applescript');
+const { runScript, renderScript } = require('../../helpers/macos/applescript');
+const { parseCodePoints } = require('../../helpers/macos/parseCodePoints');
 
 /** @type {ATDriverModules.InteractionPressKeys} */
 async function pressKeys(websocket, { keys }) {
