+++ lib/modules/session.js
@@ -5,7 +5,7 @@
 const { loadOsModule } = require('../helpers/load-os-module');
 
 /** @type {ATDriverModules.Session} */
-module.exports = loadOsModule({
+module.exports = loadOsModule('session', {
   win32() {
     return require('./win32/session');
   },
